Code quality improvements.
Improved the quality of the code base, in my opinion, substantially. Simplified and refactored the methods which create the basic NFAs of Thompson's construction. Refactored the NFA class and created a factory method which creates a NFA instance from json description. Finally, deleted the old implementation of NFAs and DFAs and the tests associated with it.